fix: update AVCaptureSession.Preset resolution, enable ultrawide camera
iPhone devices after iPhone 13 have macro mode of capturing, in order for device to use it ultraWide angle must be enabled. this commit enables ultraWide(tripleCamera) if the device supports it, if not it enables dualCamera, if none of them is available then it uses wideCamera that every iPhone has.

also this commit changes AVCaptureSession.Preset resolution to hd1920x1080, since it is not anymore problem for the device to analyze frames of that size. It improves oveerall scanning, scanning is much faster.